+++ R/custom_tidiers.R
@@ -140,13 +140,22 @@ tidy_with_broom_or_parameters <- function(x, conf.int = TRUE, conf.level = .95, 
     return(tidy_vgam(x, conf.int = conf.int, conf.level = conf.level, ...))
   }
 
-  if (inherits(x, "svy_vglm")) {
-    cli::cli_alert_info("{.cls svy_vglm} model detected.")
-    cli::cli_alert_success("{.fn tidy_svy_vglm} used instead.")
+  if (inherits(x, "coxphms")) {
+    cli::cli_alert_info("{.cls coxphms} model detected.")
+    cli::cli_alert_success("{.fn tidy_coxphms} used instead.")
     cli::cli_alert_info(
-      "Add {.code tidy_fun = broom.helpers::tidy_svy_vglm} to quiet these messages."
+      "Add {.code tidy_fun = broom.helpers::tidy_coxphms} to quiet these messages."
     )
-    return(tidy_svy_vglm(x, conf.int = conf.int, conf.level = conf.level, ...))
+    return(tidy_coxphms(x, conf.int = conf.int, conf.level = conf.level, ...))
+  }
+
+  if (inherits(x, "vglm")) {
+    cli::cli_alert_info("{.cls vglm} model detected.")
+    cli::cli_alert_success("{.fn tidy_vgam} used instead.")
+    cli::cli_alert_info(
+      "Add {.code tidy_fun = broom.helpers::tidy_vgam} to quiet these messages."
+    )
+    return(tidy_vgam(x, conf.int = conf.int, conf.level = conf.level, ...))
   }
 
   tidy_args <- list(...)
@@ -607,4 +616,65 @@ tidy_svy_vglm <- function(
   )
 
   .process_vgam_tidy_tbl(res, x$fit)
+}
+
+
+#' Tidy a multi-state survival model
+#'
+#' `r lifecycle::badge("experimental")`
+#' A tidier for multi-state models (`survival::coxphms.object`) generated
+#' with `survival::coxph()`.
+#' Term names will be updated to be consistent with generic models. The original
+#' term names are preserved in an `"original_term"` column. An additional column
+#' `"state"` will be added to provide status detail (i.e. the right part of
+#' original terms) and a column `"y.level"` will be populated by using values
+#' stored in `x$states`.
+#' @param x (`coxphms`)\cr
+#' A `survival::coxphms.object` model.
+#' @param conf.int (`logical`)\cr
+#' Whether or not to include a confidence interval in the tidied output.
+#' @param conf.level (`numeric`)\cr
+#' The confidence level to use for the confidence interval
+#' (between `0` ans `1`).
+#' @param ... Additional parameters passed to `parameters::model_parameters()`.
+#' @export
+#' @family custom_tidiers
+tidy_coxphms <- function(
+    x,
+    conf.int = TRUE,
+    conf.level = .95,
+    ...) {
+  if (!inherits(x, "coxphms")) {
+    cli::cli_abort("{.arg x} should be of class {.cls coxphms}.")
+  } # nolint
+
+  res <- tidy_parameters(
+    x,
+    conf.int = conf.int,
+    conf.level = conf.level,
+    ...
+  )
+
+  res <-
+    res |>
+    dplyr::mutate(
+      original_term = .data$term
+    ) |>
+    tidyr::separate(
+      col = dplyr::all_of("original_term"),
+      into = c("term", "state"),
+      sep = "_",
+      remove = FALSE
+    ) |>
+    tidyr::separate(
+      col = dplyr::all_of("state"),
+      into = c(NA, "y.level"),
+      sep = ":",
+      remove = FALSE
+    ) |>
+    dplyr::mutate(
+      y.level = x$states[as.integer(.data$y.level)]
+    )
+
+  res
 }